polish (moved classes from types, aded Java.hs files)
