+++ Language/Java/Java.hs
@@ -1,0 +1,10 @@
+module Language.Java.Java (
+    module Language.Java.Syntax,
+    module Language.Java.SyntaxClasses,
+    module Language.Java.Syntax.Types,
+    module Language.Java.Syntax.Exp
+) where
+import           Language.Java.Syntax
+import           Language.Java.Syntax.Exp
+import           Language.Java.Syntax.Types
+import           Language.Java.SyntaxClasses
