+++ Language/Java/Lenses/Java.hs
@@ -1,5 +1,8 @@
-module Language.Java.Lenses.Lenses
+module Language.Java.Lenses.Java
     ( module Language.Java.Syntax
+    , module Language.Java.Syntax.Types
+    , module Language.Java.Syntax.Exp
+    , module Language.Java.SyntaxClasses
     , module Language.Java.Lenses.Syntax
     , module Language.Java.Lenses.Syntax.Types
     --, module Language.Java.Lenses.Syntax.Exp
@@ -9,3 +12,7 @@ import           Language.Java.Lenses.Syntax
 --import           Language.Java.Prisms.Syntax.Exp
 import           Language.Java.Lenses.Syntax.Types
 import           Language.Java.Syntax              ()
+import           Language.Java.Syntax.Exp
+import           Language.Java.Syntax.Types        ()
+import           Language.Java.Syntax.Types        ()
+import           Language.Java.SyntaxClasses
